refactor(app): extract HardwareScan to runtime/hwscan.py (#79)
Move the 10-method hardware discovery/probe cluster out of Runtime into
a new HardwareScan collaborator (src/vanchor/runtime/hwscan.py):
  hw_scan, hw_probe, _hw_probe_serial, _hw_probe_i2c,
  _compass_sources, _gps_sources, _ports_in_use, _i2c_addrs_in_use,
  _driver_menus, _driver_context, _build_split_channel

HardwareScan holds self._rt back-ref (same pattern as HardwareGlue,
DeviceManager, etc.). _hw_probe_lock moves from Runtime onto HardwareScan
(only used by probe methods); a @property shim on Runtime forwards
rt._hw_probe_lock so the test suite keeps working.

Runtime retains shims for every externally-called method (server.py +
tests call hw_scan, hw_probe, _ports_in_use, _i2c_addrs_in_use,
_driver_menus, _compass_sources, _gps_sources, _build_split_channel,
_driver_context). _hwscan is constructed before _safety and _devices so
DeviceManager._construct_devices and SafetyRuntime can reach the shims
at build time.

app.py: -341 lines / +31 lines (net ~310 lines removed).
All 2137 tests pass; ruff clean.
